backend/ecommerce/src/main/java/com/publicissapient/ecommerce/controller/CategoryController.java
package com.publicissapient.ecommerce.controller;

import com.publicissapient.ecommerce.entities.Category;
import com.publicissapient.ecommerce.service.impl.CategoryServiceImpl;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.stream.Collectors;

@CrossOrigin
@RestController
@RequestMapping("api/categories")
public class CategoryController {
    private final CategoryServiceImpl categoryService;

    @Autowired
    public CategoryController(CategoryServiceImpl categoryService) {
        this.categoryService = categoryService;
    }

    @PostMapping("/addCategory")
    public ResponseEntity<Category> addCategory(@RequestBody Category category) {
        Category addedCategory = categoryService.addCategory(category);
        return new ResponseEntity<>(addedCategory, HttpStatus.CREATED);
    }

    @GetMapping("/getAllCategoryNames")
    public ResponseEntity<List<String>> getAllCategoryNames() {
        List<Category> categories = categoryService.getAllCategories();
        List<String> categoryNames = categories.stream()
                .map(Category::getName)
                .collect(Collectors.toList());
        return new ResponseEntity<>(categoryNames, HttpStatus.OK);
    }


}
